Solves display errors on Mac mini M1 (Monterey)
When switching tabs, sometimes the content of the SubscriptionFrame and the PublishHistoryFrame are not displayed completely. This workaround triggers a refresh of the display. If necessary, the current delay (50 ms) can be adjusted for other systems.
BTW: I love MQTTk. Finally a perfect replacement for MQTT.fx. Great work! Looking forward to the further development.